requirements_local.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0
python-decouple==3.8
redis==7.1.0
service-identity==24.2.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
zope.interface==8.1.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Azure-specific packages
django-redis==5.4.0
django-storages[azure]==1.14.2
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# Auth
firebase-admin==7.1.0
PyJWT==2.10.1

# Google API
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Payments
requests==2.31.0
google-generativeai==0.8.3

# Markdown and HTML sanitization
bleach==6.1.0
markdown==3.7
